fix(llm_eval): cast np.mean to float for axvline (pre-existing type-check debt)
Unblocks rcabench-platform 'just ci' (pyright) which this PR is the first to
trigger in a while — report.py:489 passed a numpy floating[Any] where axvline
wants a float. Unrelated to the detector change; fixed to get CI green.

+++ rcabench-platform/src/rcabench_platform/v3/sdk/llm_eval/eval/report.py
@@ -486,7 +486,7 @@ class EvalReportGenerator:
         ax.set_xlabel("F1 Score")
         ax.set_ylabel("Frequency")
         ax.set_title("F1 Score Distribution")
-        ax.axvline(np.mean(f1_values), color="#e74c3c", linestyle="--", label=f"Mean={np.mean(f1_values):.3f}")
+        ax.axvline(float(np.mean(f1_values)), color="#e74c3c", linestyle="--", label=f"Mean={np.mean(f1_values):.3f}")
         ax.legend()
 
         chart_path = self.output_dir / "f1_distribution.png"
